requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
mysql-connector-python==9.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
